Move Kubernetes specific http helpers to lib/kube
`auth.ImpersonatorRoundTripper` and `auth.IdentityForwardingHeaders`
have been relocated to lib/kube/internal. Both types were only
consumed within lib/kube and were the last remaining reason that
lib/kube depended on lib/auth.
